wire: check that final total account value is same as starting total value
Unlike `NoOverdrafts`, this is a temporal property. Temporal properties
check that every possible "lifetime" of the algorithm, from start to
finish, obeys something that relates different states in the sequence to
each other. It is similar to checking eventual consistency in databases.

`<>[]` is the "eventually-always" operator, and means that no matter what
the algorithm does, in the end the given equation must eventually be true.
It may be false for a while, and it may switch between true and false
several times, but it will end true.

Temporal property violation in this case:

```
Error: Temporal properties were violated.
Error: The following behavior constitutes a counter-example:

State 1: <Initial predicate>
/\ people = {"alice", "bob"}
/\ sender = <<"alice", "alice">>
/\ receiver = <<"bob", "bob">>
/\ acc = [alice |-> 5, bob |-> 5]
/\ amount = <<1, 1>>
/\ pc = <<"CheckAndWithdraw", "CheckAndWithdraw">>

State 2: <CheckAndWithdraw line 52, col 27 to line 58, col 77 of module wire>
/\ people = {"alice", "bob"}
/\ sender = <<"alice", "alice">>
/\ receiver = <<"bob", "bob">>
/\ acc = [alice |-> 4, bob |-> 5]
/\ amount = <<1, 1>>
/\ pc = <<"CheckAndWithdraw", "Deposit">>

State 3: <CheckAndWithdraw line 52, col 27 to line 58, col 77 of module wire>
/\ people = {"alice", "bob"}
/\ sender = <<"alice", "alice">>
/\ receiver = <<"bob", "bob">>
/\ acc = [alice |-> 3, bob |-> 5]
/\ amount = <<1, 1>>
/\ pc = <<"Deposit", "Deposit">>

State 4: <Deposit line 60, col 18 to line 63, col 68 of module wire>
/\ people = {"alice", "bob"}
/\ sender = <<"alice", "alice">>
/\ receiver = <<"bob", "bob">>
/\ acc = [alice |-> 3, bob |-> 6]
/\ amount = <<1, 1>>
/\ pc = <<"Done", "Deposit">>

State 5: Stuttering
Finished checking temporal properties in 00s at 2025-02-02 20:21:28
332 states generated, 222 distinct states found, 0 states left on queue.
Finished in 00s at (2025-02-02 20:21:28)
```

`Stuttering` is when the process simply stops, say due to power outages
or server crashes.

To fix it, we could relax the `NoOverdrafts` requirement. In the end,
this is how most banks do it: instead of guaranteeing that overdrafts
never happen, they work to make overdrafts less likely and have
procedures in place when they do happen.

Specifying safe transfers is surprisingly hard!